requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
boto3==1.38.42
boto3-stubs==1.38.46
botocore==1.38.42
botocore-stubs==1.38.30
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.4
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.9
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.23.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.2
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hiredis==3.2.1
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.2
mypy==1.16.1
mypy-boto3-cloudformation==1.38.31
mypy-boto3-dynamodb==1.38.4
mypy-boto3-ec2==1.38.45
mypy-boto3-lambda==1.38.40
mypy-boto3-rds==1.38.46
mypy-boto3-s3==1.38.44
mypy-boto3-sqs==1.38.0
mypy_extensions==1.1.0
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.0
opencv-python-headless==4.11.0.86
orjson==3.11.2
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.14
piexif==1.1.3
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
postgrest==1.1.1
pre_commit==4.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-mock==3.14.1
pytest-mypy-plugins==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pyvips==3.0.0
PyYAML==6.0.2
realtime==2.5.3
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
resend==2.13.1
rpds-py==0.25.1
rq==2.4.0
rsa==4.9.1
ruff==0.12.1
s3transfer==0.13.0
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==2.30.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.46.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.14.0
tenacity==8.5.0
testcontainers==4.13.0
tifffile==2025.6.11
tomlkit==0.13.3
tqdm==4.67.1
types-awscrt==0.27.2
types-cffi==1.17.0.20250523
types-pyOpenSSL==24.1.0.20240722
types-python-dateutil==2.9.0.20250516
types-redis==4.6.0.20241004
types-s3transfer==0.13.0
types-setuptools==80.9.0.20250529
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
